requirements.txt
requests==2.32.4
pillow==11.2.1
